[379083] Contributes babel-messages editor enhancements and bug fixes
= org.eclipse.babel.core: =
Introduced central object management system (core-class: RBManager). This
functionality was offered by TapiJIs' ResourceBundleManager
class, but due to this merge (Babel & TapiJI), I considered to move the
object management to the org.eclipse.babel.core plug-in. That plug-in
contains of the model and therefore
it's a good idea to place object management there too. To ensure loose
coupling, I defined interfaces, which are widley used by TapiJI.
Due to the merge of these two worlds, some other functionalities had to be
moved from one to another (mostly from TapiJI to Babel). One example is
the ConfigurationManager. This class has access to the preferences (defined in
TapiJI) via the babelConfiguration Xpt. Some factory classes had to be
introduced as well.
An example is the MessagesBundleGroupFactory class. It creates bundle
groups based on a PropertiesFileGroupStrategy, which is quite uncommon,
but in some cases of vital importance.
Also some Utility-classes had to be implemented.

= org.eclipse.babel.editor: =
Modifications were made due to huge changes in org.eclipse.babel.core
(writing Interfaces, ...). Also many bugs have been fixed or partly fixed
(only 222152) during the integration with TapiJI.
(see: Bugs: 222127, 222131, 222138, 222152, 222140, 226383, 226389,
226390). I also had to implement workarounds for special problems, see
class DirtyHack.

[222127] fix [222131] fix [222138] fix [222140] fix [226383] fix
[226389] fix

close 379083