[379084] Contributes I18N coding assistances
Contributes the following plug-ins for extending the Eclipse Java IDE with smart tools for reducing the effort of Internationalization.

* `org.eclipse.babel.tapiji.tools.core`: Core plug-in for I18N tooling
* `org.eclipse.babel.tapiji.tools.core.ui`: Common UI elements for I18N tooling
* `org.eclipse.babel.tapiji.tools.java`: Java programming language specific I18N assistances
* `org.eclipse.babel.tapiji.tools.java.ui`: UI elements for Java I18N

Internationalization based on the basic Java API is a tedious task and introduces additional complexity into development. For example, typical faults are:

* The export of culture-neutral configuration strings
* The incomplete export of constant string literals
* To hard code constant string literals
* To build static format strings and
* To reuse externalized literals in different contexts

These problems are addressed by the contributed tooling with the goal of making I18N tasks more transparent. For example, hard coded cultural aspects as well as broken Internationalizations are reported to the problem view and can be immediately resolved with a rich set of quick fixes.

close 379084